Give recordTake & editTake implementations and dialog properties to RecordableViewModel
