+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/resourcetakes/viewmodel/TabRecordableViewModel.kt
@@ -4,16 +4,15 @@ import javafx.beans.binding.Bindings
 import javafx.beans.binding.StringBinding
 import javafx.beans.property.SimpleStringProperty
 import org.wycliffeassociates.otter.common.data.workbook.Take
+import org.wycliffeassociates.otter.jvm.app.ui.takemanagement.viewmodel.AudioPluginViewModel
 import org.wycliffeassociates.otter.jvm.app.ui.takemanagement.viewmodel.RecordableViewModel
 import java.util.concurrent.Callable
 
 class TabRecordableViewModel(
     val labelProperty: SimpleStringProperty,
-    recordNewTake: () -> Unit,
-    editTake: (Take) -> Unit
+    audioPluginViewModel: AudioPluginViewModel
 ): RecordableViewModel(
-    recordNewTake,
-    editTake
+    audioPluginViewModel
 ) {
     fun getFormattedTextBinding(): StringBinding = Bindings.createStringBinding(
         Callable { getFormattedText() },
