+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/widgets/takecard/ScriptureTakeCardSkin.kt
@@ -35,7 +35,7 @@ class ScriptureTakeCardSkin(control: TakeCard) : TakeCardSkin(control) {
                             ) {
                                 addClass(TakeCardStyles.takeNumberLabel)
                             }
-                            label(control.take.modifiedTimestamp.value.toString()) {
+                            label(control.take.createdTimestamp.toString()) {
                                 addClass(TakeCardStyles.timestampLabel)
                             }
                         }
